+++ packages/client/src/components/Select/Select.tsx
@@ -1,71 +1,16 @@
-/* eslint-disable @typescript-eslint/no-unused-expressions */
-/* eslint-disable jsx-a11y/interactive-supports-focus */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-
 import { useState, useEffect, useRef } from 'react'
 import type { MouseEventHandler } from 'react'
+import type { TOption } from '@/pages/leader-board/leader-board'
 
 import { TLeaderboardPlayerData } from '@/models/leaderboard'
 import { ReactComponent as ArrowDown } from '@/assets/arrow-down.svg'
+import { Option } from './Option/Option'
 
 import styles from './Select.module.scss'
 
-type Option = {
-  title: string
-  value: keyof Omit<TLeaderboardPlayerData, 'id'>
-}
-
-type OptionProps = {
-  option: Option
-  onClick: (value: keyof Omit<TLeaderboardPlayerData, 'id'>) => void
-}
-
-const OptionEl = (props: OptionProps) => {
-  const {
-    option: { value, title },
-    onClick,
-  } = props
-
-  const optionRef = useRef<HTMLLIElement>(null)
-
-  const handleClick =
-    (
-      clickedValue: keyof Omit<TLeaderboardPlayerData, 'id'>
-    ): MouseEventHandler<HTMLLIElement> =>
-    () => {
-      onClick(clickedValue)
-    }
-
-  useEffect(() => {
-    const option = optionRef.current
-    if (!option) return
-    const handleEnterKeyDown = (event: KeyboardEvent) => {
-      if (document.activeElement === option && event.key === 'Enter') {
-        onClick(value)
-      }
-    }
-
-    option.addEventListener('keydown', handleEnterKeyDown)
-    return () => {
-      option.removeEventListener('keydown', handleEnterKeyDown)
-    }
-  }, [value, onClick])
-
-  return (
-    <li
-      className={styles.option}
-      value={value}
-      onClick={handleClick(value)}
-      ref={optionRef}>
-      {title}
-    </li>
-  )
-}
-
 type SelectProps = {
-  selected: Option | null
-  options: Option[]
+  selected: TOption | null
+  options: TOption[]
   placeholder?: string
   onChange?: (selected: keyof Omit<TLeaderboardPlayerData, 'id'>) => void
   onClose?: () => void
@@ -82,7 +27,9 @@ export const Select = (props: SelectProps) => {
     const handleClick = (event: MouseEvent) => {
       const { target } = event
       if (target instanceof Node && !rootRef.current?.contains(target)) {
-        isOpen && onClose?.()
+        if (isOpen && onClose) {
+          onClose()
+        }
         setIsOpen(false)
       }
     }
@@ -137,7 +84,7 @@ export const Select = (props: SelectProps) => {
       {isOpen && (
         <ul className={styles.select}>
           {options.map(option => (
-            <OptionEl
+            <Option
               key={option.value}
               option={option}
               onClick={handleOptionClick}
